A-Packages: Add packages for project
