+++ src/main/dobbyList.txt
@@ -6,3 +6,5 @@ E | [ ] | test1 | Feb 02 2022 00:00
 D | [ ] | test2 | Jan 01 2021 02:02
 T | [ ] | test3
 T | [ ] | test1
+T | [ ] | test
+T | [ ] | chunkhai
