+++ src/main/java/Dobby/commands/ByeCommand.java
@@ -1,3 +1,8 @@
+package dobby.commands;
+
+import dobby.tasks.*;
+import dobby.*;
+
 import java.io.IOException;
 
 public class ByeCommand extends Command{
