+++ src/main/java/Dobby/commands/Command.java
@@ -1,3 +1,8 @@
+package dobby.commands;
+
+import dobby.tasks.*;
+import dobby.*;
+
 import java.io.IOException;
 
 public abstract class Command {
