+++ src/main/java/Dobby/commands/DeleteCommand.java
@@ -1,3 +1,8 @@
+package dobby.commands;
+
+import dobby.tasks.*;
+import dobby.*;
+
 public class DeleteCommand extends Command{
     @Override
     public void execute(DobbyList dl, UserInput ui) {
