+++ src/main/java/Dobby/commands/ErrorCommand.java
@@ -1,3 +1,8 @@
+package dobby.commands;
+
+import dobby.tasks.*;
+import dobby.*;
+
 public class ErrorCommand extends Command{
     @Override
     public void execute(DobbyList dl, UserInput ui) {
